requirements.txt
google-generativeai
numpy
nest_asyncio
uvicorn
fastapi
python-telegram-bot
requests
pydantic
faiss-cpu